// sdk/src/models/scheduled-message-ref.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import { safeParse } from "../lib/schemas.js";
import { Result as SafeParseResult } from "../types/fp.js";
import * as types from "../types/primitives.js";
import { SDKValidationError } from "./errors/sdk-validation-error.js";

/**
 * Lightweight reference returned by `createScheduledMessage`.
 */
export type ScheduledMessageRef = {
  id: string;
  channelId: string;
};

/** @internal */
export const ScheduledMessageRef$inboundSchema: z.ZodType<
  ScheduledMessageRef,
  z.ZodTypeDef,
  unknown
> = z.object({
  id: types.string(),
  channelId: types.string(),
});
/** @internal */
export type ScheduledMessageRef$Outbound = {
  id: string;
  channelId: string;
};

/** @internal */
export const ScheduledMessageRef$outboundSchema: z.ZodType<
  ScheduledMessageRef$Outbound,
  z.ZodTypeDef,
  ScheduledMessageRef
> = z.object({
  id: z.string(),
  channelId: z.string(),
});

export function scheduledMessageRefToJSON(
  scheduledMessageRef: ScheduledMessageRef,
): string {
  return JSON.stringify(
    ScheduledMessageRef$outboundSchema.parse(scheduledMessageRef),
  );
}
export function scheduledMessageRefFromJSON(
  jsonString: string,
): SafeParseResult<ScheduledMessageRef, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => ScheduledMessageRef$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'ScheduledMessageRef' from JSON`,
  );
}